requirements.txt
# RAG System Dependencies

Flask==2.3.0
Flask-CORS==4.0.0
PyMuPDF==1.23.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
anthropic==0.18.0
numpy==1.24.0
pandas==2.0.0
python-dotenv==1.0.0
requests==2.31.0
